+++ src/infrastructure/improvement.h
@@ -144,7 +144,7 @@ public:
 		return this->wealth_cost;
 	}
 
-	const commodity_map<int> &get_commodity_costs() const
+	const commodity_map<int64_t> &get_commodity_costs() const
 	{
 		return this->commodity_costs;
 	}
@@ -190,7 +190,7 @@ private:
 	std::vector<const improvement *> requiring_improvements; //improvements which require this one
 	technology *required_technology = nullptr;
 	int wealth_cost = 0;
-	commodity_map<int> commodity_costs;
+	commodity_map<int64_t> commodity_costs;
 	std::unique_ptr<const and_condition<site>> conditions;
 	std::unique_ptr<const and_condition<site>> free_on_start_conditions;
 	std::unique_ptr<metternich::modifier<const site>> modifier;
